refactor: split crud tests
